Testing new way to run batch mode
Uses sequential steps rather than all-in-one. Default method relied on opeing file, processing prediction, saving results. This method opens all files and makes an np.array, processes the array, savest the results. The performance is inferior at the moment. (180 vs 120 secs) but possibly because of the double use of for loops for processing np.arrays.