Experimental: test embedding input_shape into arch
Input_shape is added before the definition fo the architecture. When
keras saves the model it excludes the input_shape, which is needed for
further converisons to JS models. Moving the input_shape definition in
the definition fo the first layer of the achitecture should fix that.
This is an experimental version for testing. If successful it will be
moved to stable.